patient data set
K nearest neighbors is a simple supervised learning algorithm that stores all available cases and classifies new cases based on a similarity measure (e.g., distance functions).
In this particular data set ,after performing training on data set by KNN classification , patient is predicted to be affected or not. A confusion matrix is created based on correct  and false prediction along with the accuracy score. This classification is done for both two and three class data.